Creation of NoStateWorker and StateWorker classes
